requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
boto3==1.34.44
botocore==1.34.44
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
easyocr==1.7.1
fastapi==0.109.2
filelock==3.13.1
Flask==3.0.2
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
langchainhub==0.1.14
langsmith==0.0.87
lazy_loader==0.3
MarkupSafe==2.1.5
marshmallow==3.20.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
openai==1.12.0
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
pinecone-client==3.0.2
psycopg2-binary==2.9.9
pyclipper==1.3.0.post5
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tifffile==2024.2.12
tiktoken==0.5.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.27.1
Werkzeug==3.0.1
yarl==1.9.4
